Rename src/pkg/[library/shub]/client -> src/pkg/client/[library/shub]
