+++ src/cmd/singularity/cli/search.go
@@ -7,7 +7,7 @@ package cli
 
 import (
 	"github.com/singularityware/singularity/src/docs"
-	"github.com/singularityware/singularity/src/pkg/library/client"
+	"github.com/singularityware/singularity/src/pkg/client/library"
 	"github.com/singularityware/singularity/src/pkg/sylog"
 	"github.com/spf13/cobra"
 )
@@ -26,8 +26,8 @@ func init() {
 // SearchCmd singularity search
 var SearchCmd = &cobra.Command{
 	DisableFlagsInUseLine: true,
-	Args:   cobra.ExactArgs(1),
-	PreRun: sylabsToken,
+	Args:                  cobra.ExactArgs(1),
+	PreRun:                sylabsToken,
 	Run: func(cmd *cobra.Command, args []string) {
 		if err := client.SearchLibrary(args[0], SearchLibraryURI, authToken); err != nil {
 			sylog.Fatalf("Couldn't search library: %v", err)
