+++ src/pkg/client/library/search.go
