+++ src/pkg/client/library/search_test.go
